Initial go at a capnproto OpenBSD port
The only really horrible thing in this is that I couldn't figure out how
to get OpenBSD's SHARED_LIBS functionality to work with -release, so I
killed it and replaced it with a -version-info. I definitely don't fully
understand the consequences of this, but maybe it's good enough to work
with.